+++ server/config/default.ts
@@ -1,0 +1,5 @@
+export default {
+  port: 5000,
+  host: "localhost",
+  dbUri: "mongodb+srv://nap-admin:gzRpnLItCZaAmiix@cluster0.owcgb.mongodb.net/rick-and-morty?retryWrites=true&w=majority"
+}